compat: give a glob's counts the widths POSIX gives them
The glob_t the Windows build supplies counted its matches in an int
where POSIX counts them in a size_t. A caller written against the type
POSIX defines then compared a cursor with the count in whichever
signedness the platform happened to give it, which the compiler refuses
outright where warnings are errors: the tree did not build.

The structure now carries the widths POSIX gives it, so one piece of
code reads one glob_t whichever platform supplied it, and the count no
longer passes through a narrowing cast on its way in.

Commit message generated by Claude (Anthropic).
Source code generated by Claude (Anthropic).

+++ src/lib/xpost_compat.c
@@ -304,7 +304,7 @@ xpost_glob(const char *pattern, glob_t *pglob)
                 }
 
                 *begin = NULL;
-                pglob->gl_pathc = (int)matches;
+                pglob->gl_pathc = matches;
                 pglob->gl_pathv = (char**)new_buffer;
             }
         }
@@ -315,7 +315,7 @@ xpost_glob(const char *pattern, glob_t *pglob)
 
     if (result == 0)
     {
-        if ((size_t)pglob->gl_pathc == max_matches)
+        if (pglob->gl_pathc == max_matches)
             result = -1;
     }
